requirements.txt
streamlit==1.31.0
langchain==0.1.3
langchain-openai==0.0.3
python-dotenv==1.0.0
openai==1.10.0
torch==2.1.2
chromadb==0.3.29
tiktoken==0.5.2
pypdf==3.17.3
